Add wait logic to dev/prod docker environments
Noticed failing CI jobs because postgres was not spinning up fast enough
for django. Brought in an external wait script, thanks vishnubob/wait-for-it.

+++ docker/django-start.sh
@@ -2,17 +2,32 @@
 #
 
 wait_for_node() {
-    while [ ! -f .node_complete ]
-    do
-        sleep 2
-    done
-    rm -v .node_complete
+    if [ "${DEPLOY_ENV}" != "prod" ]; then
+        echo Waiting for node to start..
+        while [ ! -f .node_complete ]
+        do
+            sleep 2
+        done
+        rm -v .node_complete
+    fi
+}
+
+wait_for_postgres() {
+    echo Waiting for postgres to start..
+    wait-for-it.sh -t 120 -h ${DJANGO_DB_HOST} -p ${DJANGO_DB_PORT}
 }
 
 django_start() {
-    ./manage.py migrate && ./manage.py runserver 0.0.0.0:8000
+    if [ "${DEPLOY_ENV}" == "prod" ]; then
+        ./manage.py migrate
+        ./manage.py collectstatic -c --noinput
+        gunicorn -c /etc/gunicorn/gunicorn.py securethenews.wsgi
+    else
+        ./manage.py migrate && ./manage.py runserver 0.0.0.0:8000
+    fi
 }
 
 
+wait_for_postgres
 wait_for_node
 django_start
